extract monarch related files to a folder
