+++ package/src/syntax-highlighting/kustoMonarchLanguageDefinition.ts
